+++ crates/glyph/src/pane/gutter/relative_line_gutter.rs
@@ -33,7 +33,7 @@ impl Gutter for RelativeLineDrawer {
                 }
             }
 
-            line = " ".repeat(self.config.gutter_width as usize - 1 - line.len()) + &line;
+            line = " ".repeat(self.config.gutter_width - 1 - line.len()) + &line;
             line.push(' ');
 
             for (x, c) in line.chars().enumerate() {
@@ -42,7 +42,7 @@ impl Gutter for RelativeLineDrawer {
         }
 
         if total_lines < viewport.height {
-            let mut line = " ".repeat(self.config.gutter_width as usize - 2);
+            let mut line = " ".repeat(self.config.gutter_width - 2);
             line.push(self.config.empty_line_char);
             line.push(' ');
